requirements.txt
Django==1.11.20
django-extensions==2.0.7
psycopg2-binary==2.7.4
unittest-xml-reporting==1.11.0
django-govuk-template==0.4
django-govuk-forms==0.4
requests==2.18.4
whitenoise==3.3.1
inflect==0.2.5
django-debug-toolbar==1.9.1
django-timeline-logger==1.0
selenium==3.11.0
faker==0.8.13
hypothesis==3.70.3
django-xhtml2pdf==0.0.4
boto3==1.7.66
lxml==4.3.0
